+++ src/test/java/com/arangodb/example/ssl/SslExampleTest.java
@@ -18,34 +18,34 @@
  * Copyright holder is ArangoDB GmbH, Cologne, Germany
  */
 
-package com.arangodb.async;
+package com.arangodb.example.ssl;
 
-import com.arangodb.ArangoDBException;
+import com.arangodb.ArangoDB;
+import com.arangodb.Protocol;
 import com.arangodb.entity.ArangoDBVersion;
-import org.junit.Ignore;
-import org.junit.Test;
+import org.apache.http.conn.ssl.NoopHostnameVerifier;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
 import javax.net.ssl.KeyManagerFactory;
 import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLHandshakeException;
 import javax.net.ssl.TrustManagerFactory;
 import java.security.KeyStore;
 
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.notNullValue;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.fail;
+import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author Mark Vollmary
+ * @author Michele Rastelli
  */
-public class ArangoSslTest {
+@EnabledIfSystemProperty(named = "SslTest", matches = "true")
+class SslExampleTest {
 
     /*-
      * a SSL trust store
      *
      * create the trust store for the self signed certificate:
-     * keytool -import -alias "my arangodb server cert" -file UnitTests/server.pem -keystore example.truststore
+     * keytool -import -alias "my arangodb server cert" -file server.pem -keystore example.truststore
      *
      * Documentation:
      * https://hc.apache.org/httpcomponents-client-ga/httpclient/apidocs/org/apache/http/conn/ssl/SSLSocketFactory.html
@@ -54,8 +54,35 @@ public class ArangoSslTest {
     private static final String SSL_TRUSTSTORE_PASSWORD = "12345678";
 
     @Test
-    @Ignore
-    public void connect() throws Exception {
+    void connect() throws Exception {
+        final ArangoDB arangoDB = new ArangoDB.Builder()
+                .host("localhost", 8529)
+                .password("test")
+                .useSsl(true)
+                .sslContext(createSslContext())
+                .useProtocol(Protocol.HTTP_JSON)
+                .build();
+        final ArangoDBVersion version = arangoDB.getVersion();
+        assertThat(version).isNotNull();
+        System.out.println(version.getVersion());
+    }
+
+    @Test
+    void noopHostnameVerifier() throws Exception {
+        final ArangoDB arangoDB = new ArangoDB.Builder()
+                .host("127.0.0.1", 8529)
+                .password("test")
+                .useSsl(true)
+                .sslContext(createSslContext())
+                .hostnameVerifier(NoopHostnameVerifier.INSTANCE)
+                .useProtocol(Protocol.HTTP_JSON)
+                .build();
+        final ArangoDBVersion version = arangoDB.getVersion();
+        assertThat(version).isNotNull();
+        System.out.println(version.getVersion());
+    }
+
+    private SSLContext createSslContext() throws Exception {
         final KeyStore ks = KeyStore.getInstance(KeyStore.getDefaultType());
         ks.load(this.getClass().getResourceAsStream(SSL_TRUSTSTORE), SSL_TRUSTSTORE_PASSWORD.toCharArray());
 
@@ -68,25 +95,7 @@ public class ArangoSslTest {
         final SSLContext sc = SSLContext.getInstance("TLS");
         sc.init(kmf.getKeyManagers(), tmf.getTrustManagers(), null);
 
-        final ArangoDBAsync arangoDB = new ArangoDBAsync.Builder()
-                .loadProperties(ArangoSslTest.class.getResourceAsStream("/arangodb-ssl.properties")).useSsl(true)
-                .sslContext(sc).build();
-        final ArangoDBVersion version = arangoDB.getVersion().get();
-        assertThat(version, is(notNullValue()));
-    }
-
-    @Test
-    @Ignore
-    public void connectWithoutValidSslContext() throws Exception {
-        try {
-            final ArangoDBAsync arangoDB = new ArangoDBAsync.Builder()
-                    .loadProperties(ArangoSslTest.class.getResourceAsStream("/arangodb-ssl.properties")).useSsl(true)
-                    .build();
-            arangoDB.getVersion().get();
-            fail("this should fail");
-        } catch (final ArangoDBException ex) {
-            assertThat(ex.getCause() instanceof SSLHandshakeException, is(true));
-        }
+        return sc;
     }
 
 }
